Final/fox_detection/requirements.txt
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
Pillow>=9.0.0
tqdm>=4.64.0
jupyter>=1.0.0
gradio>=4.0.0
